api/requirements.txt
# Core Flask dependencies
Flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-limiter==3.8.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
alembic==1.13.2

# Authentication & Security
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation
pydantic==2.9.2

# Celery & Redis
celery==5.4.0
redis==5.1.1

# HTTP requests
requests==2.32.3

# Scientific computing
numpy==2.1.1

# Production WSGI server
gunicorn==23.0.0

# Development/Testing
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==5.0.0